split it_266F
