+++ config/GALE01/symbols.txt
@@ -20165,6 +20165,7 @@ it_803B857C = .rodata:0x803B857C; // type:object size:0xC scope:global data:4byt
 it_803B8588 = .rodata:0x803B8588; // type:object size:0x10 scope:global data:4byte
 it_803B8598 = .rodata:0x803B8598; // type:object size:0x10 scope:global data:4byte
 it_803B85A8 = .rodata:0x803B85A8; // type:object size:0x48 scope:global data:4byte
+...rodata.0 = .rodata:0x803B85A8; // type:label scope:local
 @541 = .rodata:0x803B85F0; // type:object size:0x10 scope:local data:4byte
 it_803B8600 = .rodata:0x803B8600; // type:object size:0xC scope:global data:4byte
 it_803B8610 = .rodata:0x803B8610; // type:object size:0x24 scope:global data:4byte
@@ -32865,8 +32866,8 @@ mnCharSel_804DC624 = .sdata2:0x804DC624; // type:object size:0x4 scope:global al
 it_804DC698 = .sdata2:0x804DC698; // type:object size:0x8 scope:global align:8 data:double
 it_804DC6A0 = .sdata2:0x804DC6A0; // type:object size:0x4 scope:global align:4 data:float
 it_804DC6A4 = .sdata2:0x804DC6A4; // type:object size:0x4 scope:global align:4 data:float
-it_804DC6A8 = .sdata2:0x804DC6A8; // type:object size:0x4 scope:global align:4 data:float
-it_804DC6AC = .sdata2:0x804DC6AC; // type:object size:0x4 scope:global align:4 data:float
+@304 = .sdata2:0x804DC6A8; // type:object size:0x4 scope:local align:4 data:float
+@305 = .sdata2:0x804DC6AC; // type:object size:0x4 scope:local align:4 data:float
 it_804DC6B0 = .sdata2:0x804DC6B0; // type:object size:0x4 scope:global align:4 data:float
 it_804DC6B8 = .sdata2:0x804DC6B8; // type:object size:0x4 scope:global align:4 data:float
 it_804DC6BC = .sdata2:0x804DC6BC; // type:object size:0x4 scope:global align:4 data:float
