+++ src/melee/it/it_266F.h
@@ -2,85 +2,9 @@
 #define GALE01_26C47C
 
 #include <placeholder.h>
-#include <platform.h>
 
-#include "it/types.h"
-
-#include <baselib/forward.h>
-
-#include <dolphin/mtx.h>
-
-/* 26C47C */ void it_8026C47C(struct it_8026C47C_arg0_t*);
-/* 26C530 */ s32 it_8026C530(s32, ItemPickTable*, s32, s32);
-/* 26C65C */ int it_8026C65C(ItemPickTable*);
-/* 26C704 */ bool it_8026C704(void);
-/* 26C75C */ ItemKind it_8026C75C(ItemPickTable*);
-/* 26C88C */ void fn_8026C88C(HSD_GObj*);
-/* 26CA4C */ void it_8026CA4C(ItemPickTable*, s32*, u64, s32, f32);
-/* 26CB3C */ bool it_8026CB3C(Vec3*);
-/* 26CB9C */ void it_8026CB9C(s32*, u64, f32);
-/* 26CD50 */ void it_8026CD50(s32*, u64, f32);
-/* 26CF04 */ void it_8026CF04(void);
-/* 26D018 */ void it_8026D018(void);
-/* 26D258 */ bool it_8026D258(Vec3*, ItemKind);
-/* 26D324 */ bool it_8026D324(ItemKind);
-/* 26D3CC */ bool it_8026D3CC(void);
-/* 26D564 */ bool it_8026D564(Item_GObj*);
-/* 26D5CC */ bool it_8026D5CC(Item_GObj*);
-/* 26D604 */ bool it_8026D604(Item_GObj*);
-/* 26D62C */ void it_8026D62C(Item_GObj*, HSD_GObjEvent);
-/* 26D6F4 */ void it_8026D6F4(Item_GObj*, HSD_GObjEvent);
-/* 26D78C */ void it_8026D78C(Item_GObj*, HSD_GObjEvent);
-/* 26D82C */ void it_8026D82C(Item_GObj*);
-/* 26D8A4 */ bool it_8026D8A4(Item_GObj*, HSD_GObjEvent);
-/* 26D938 */ void it_8026D938(Item_GObj*, HSD_GObjEvent);
-/* 26D9A0 */ bool it_8026D9A0(Item_GObj*);
-/* 26DA08 */ bool it_8026DA08(Item_GObj*);
-/* 26DA70 */ bool it_8026DA70(Item_GObj*);
-/* 26DAA8 */ s32 it_8026DAA8(Item_GObj*);
-/* 26DB40 */ bool it_8026DB40(Item_GObj*);
-/* 26DBC8 */ bool it_8026DBC8(Item_GObj*);
-/* 26DC24 */ bool it_8026DC24(Item_GObj*);
-/* 26DD5C */ bool it_8026DD5C(Item_GObj*);
-/* 26DDFC */ bool it_8026DDFC(Item_GObj*);
-/* 26DE98 */ bool it_8026DE98(Item_GObj*);
-/* 26DF34 */ bool it_8026DF34(Item_GObj*);
-/* 26DFB0 */ bool it_8026DFB0(Item_GObj*);
-/* 26E058 */ bool it_8026E058(Item_GObj*);
-/* 26E0F4 */ void it_8026E0F4(Item_GObj*);
-/* 26E15C */ void it_8026E15C(Item_GObj*, HSD_GObjEvent);
-/* 26E248 */ void it_8026E248(Item_GObj*, HSD_GObjEvent);
-/* 26E248 */ void it_8026E248(Item_GObj*, HSD_GObjEvent);
-/* 26E32C */ bool it_8026E32C(Item_GObj*, HSD_GObjEvent);
-/* 26E414 */ void it_8026E414(Item_GObj*, HSD_GObjEvent);
-/* 26E4D0 */ void it_8026E4D0(Item_GObj*, HSD_GObjEvent);
-/* 26E5A0 */ void it_8026E5A0(Item_GObj*, HSD_GObjEvent);
-/* 26E664 */ void it_8026E664(Item_GObj*, HSD_GObjEvent);
-/* 26E71C */ void it_8026E71C(Item_GObj*, HSD_GObjEvent);
-/* 26E7E0 */ void it_8026E7E0(Item_GObj*, HSD_GObjEvent);
-/* 26E8C4 */ void it_8026E8C4(Item_GObj*, HSD_GObjEvent, HSD_GObjEvent);
-/* 26E9A4 */ bool it_8026E9A4(HSD_GObj*, Vec3*, Vec3*, Vec3*);
-/* 26EA20 */ bool it_8026EA20(HSD_GObj*, Vec3*, Vec3*, Vec3*, int*,
-                              u32* flags_out);
-/* 26EA9C */ bool it_8026EA9C(HSD_GObj*, Vec3*, Vec3*, Vec3*, int*,
-                              u32* flags_out, s32);
-/* 26EB18 */ void it_8026EB18(HSD_GObj*, s32, Vec3*);
-/* 26EBC8 */ void it_8026EBC8(HSD_GObj*, u16, u8*);
-/* 26EC54 */ void it_8026EC54(HSD_GObj*, u16, u8*);
-/* 26ECE0 */ bool it_8026ECE0(Item_GObj*, u32);
-/* 26EECC */ void it_8026EECC(HSD_GObj*, int);
-/* 26F3AC */ ItemKind it_8026F3AC(void);
-/* 26F3D4 */ bool it_8026F3D4(Item_GObj*, struct it_8026F3D4_arg1_t*, s32,
-                              s32);
-/* 26F53C */ void it_8026F53C(Item_GObj*, Vec3*, bool);
-/* 26F5C8 */ Item_GObj* it_8026F5C8(Item_GObj*, ItemKind, Vec3*);
-/* 26F6BC */ Item_GObj* it_8026F6BC(Item_GObj*, Vec3*, Vec3*, bool);
-/* 26F7C8 */ Item_GObj* it_8026F7C8(Vec3*, Vec3*, bool);
-/* 26F8B4 */ bool it_8026F8B4(Item_GObj*, Vec3*, Vec3*, bool);
-/* 26F9A0 */ void it_8026F9A0(void);
-
-bool it_8026E15C_inline1(Item_GObj* gobj);
-bool it_8026E15C_inline2(Item_GObj* gobj);
-bool it_8026E248_inline(Item_GObj* gobj);
-bool it_8026E32C_inline(Item_GObj* gobj);
+#include "it/itdraw.h"
+#include "it/itdrop.h"
+#include "it/itgroundcoll.h"
+#include "it/itspawn.h"
 #endif
