+++ src/melee/it/itcoll.h
@@ -7,6 +7,7 @@
 
 #include "lb/types.h"
 
+/* 26F9A0 */ void it_8026F9A0(void);
 /* 26F9AC */ void it_8026F9AC(s32, void*, HitCapsule*, Item*, HurtCapsule*);
 /* 26FA2C */ void it_8026FA2C(Item*, HitCapsule*, s32, Item*, bool);
 /* 26FAC4 */ void it_8026FAC4(Item*, HitCapsule*, s32, void*, bool);
