+++ src/melee/it/itdraw.h
@@ -1,0 +1,16 @@
+#ifndef GALE01_ITDRAW_H
+#define GALE01_ITDRAW_H
+
+#include <platform.h>
+
+#include "it/types.h"
+
+#include <baselib/forward.h>
+
+/* 26EB18 */ void it_8026EB18(HSD_GObj*, s32, Vec3*);
+/* 26EBC8 */ void it_8026EBC8(HSD_GObj*, u16, u8*);
+/* 26EC54 */ void it_8026EC54(HSD_GObj*, u16, u8*);
+/* 26ECE0 */ u32 it_8026ECE0(Item_GObj*, u32);
+/* 26EECC */ void it_8026EECC(HSD_GObj*, int);
+
+#endif
